rename constants folder to API
