+++ project-movie-app/src/API/fetchSearchResults.ts
