+++ project-movie-app/src/API/getSearchedMovies.ts
